# du: rewrite to use (mostly) working multithreaded directory walker (WIP)