### DO NOT MERGE YET!

This is a mostly working rewrite of `du` using `crossbeam` and `structopt` (although I am strongly considering switching to `clap` by itself because I am finding it difficult to get `structopt` to do what I want).  There are a few issues mentioned in #1377, and a few features (such as processing files given on the command-line that are *not* directories) are currently broken.  I am probably going to go through and try to clean up the code some more as well.

However, at this point, I think the code is more or less ready for some feedback if anyone has any.  I am considering adding a flag (or maybe an environment variable?  `POSIXLY_CORRECT` or `GNULY_CORRECT` might work for this) that just resets `du` into (mostly) normal single-threaded behavior, sort of like the `--thread-count` flag I added.

Also, note that I included a couple minor changes to `env` in this commit (one of which is adding my name to the authors, because I refactored pretty much the entire thing).

## Benchmarks

I have some (very unscientific) benchmarks made by running the new version of `du`, the old version of `du`, GNU `du`, and `diskus` on a very large directory on a very overloaded 2012 MacBook Pro.

#### new `du`
```bash
$ time target/release/uutils du -s ~
du: cannot read directory ‘/Users/alex/Library/Preferences/KDE‘: Permission denied (os error 13)
426662684	/Users/alex
target/release/uutils du -s ~  15.49s user 168.16s system 215% cpu 1:25.16 total
```

#### old `du`
```bash
$ time target/release/uutils du -s ~
du: cannot read directory ‘/Users/alex/Library/Preferences/KDE‘: Permission denied (os error 13)
du: cannot read directory ‘/Users/alex/Desktop/coreutils/tmp/coreutils-8.24/gt-deep-1.sh.evGP/t/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k/k‘: Too many open files (os error 24)
426661732	/Users/alex
target/release/uutils du -s ~  13.29s user 110.63s system 51% cpu 4:00.89 total
```

#### GNU `du`
```bash
$ time gdu -s ~
gdu: cannot read directory '/Users/alex/Library/Preferences/KDE': Permission denied
426663172	/Users/alex
gdu -s ~  6.90s user 98.50s system 47% cpu 3:43.24 total
```

#### `diskus`
```bash
$ time diskus ~

thread '<unknown>' has overflowed its stack
fatal runtime error: stack overflow
zsh: abort      diskus ~
diskus ~  1.86s user 27.79s system 428% cpu 6.913 total
```

You'll notice that while `diskus` obviously crashed, the sizes output by the three variations of `du` differ, which is obviously an issue.  I'll try to track down what the cause is, but it may take a bit of time.